requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
APScheduler==3.10.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
exceptiongroup==1.1.0
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
iniconfig==2.0.0
multidict==6.0.4
openai==0.27.0
packaging==23.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pytest==7.2.2
python-dotenv==1.0.0
python-telegram-bot==20.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.7
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.14
yarl==1.8.2
